pyproject.toml
[project]
name = "voice-agent-researcher"
version = "0.1.0"
description = "Voice agent researcher"
requires-python = ">=3.12"
dependencies = [
    "pipecat-ai[anthropic,cartesia,daily,local-smart-turn-v3,runner,silero,tracing,webrtc]",
    "pipecat-ai-cli",
    "supabase",
    "pipecatcloud>=0.2.19",
    "opentelemetry-exporter-otlp-proto-http",
]

[dependency-groups]
dev = [
    "pyright>=1.1.404,<2",
    "ruff>=0.12.11,<1",
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I"]